pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "digitalhub"
version = "0.14.0"
description = "Python SDK for Digitalhub"
readme = "README.md"
authors = [
    { name = "Fondazione Bruno Kessler", email = "digitalhub@fbk.eu" },
    { name = "Matteo Martini", email = "mmartini@fbk.eu" }
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
keywords = ["data", "dataops", "kubernetes"]
requires-python = ">=3.9, <3.13"
dependencies = [
    "boto3",
    "GitPython>=3",
    "numpy",
    "psycopg2-binary",
    "pyarrow",
    "pydantic",
    "python-slugify",
    "PyYAML",
    "requests",
    "sqlalchemy",
]

[project.optional-dependencies]
full = [
    "mlflow",
    "pandas",
]
pandas = [
    "pandas",
]
mlflow = [
    "mlflow",
]
dev = [
    "bumpver",
    "jsonschema",
    "pytest-cov",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/scc-digitalhub/digitalhub-sdk"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.bumpver]
current_version = "0.14.0"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = false
tag             = false
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
